+++ src/main/java/com/mmp/beacon/security/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.mmp.beacon.Security.config;
+package com.mmp.beacon.security.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
