+++ src/main/java/com/mmp/beacon/security/presentation/UserPresentationController.java
@@ -1,10 +1,12 @@
-package com.mmp.beacon.Security.auth;
+package com.mmp.beacon.security.presentation;
 
+import com.mmp.beacon.security.application.command.UserApplicationService;
+import com.mmp.beacon.security.config.UserDetail;
+import com.mmp.beacon.security.presentation.request.LoginRequest;
 import com.mmp.beacon.company.domain.Company;
 import com.mmp.beacon.company.domain.repository.CompanyRepository;
-import com.mmp.beacon.Security.presentation.request.CreateUserRequest;
-import com.mmp.beacon.Security.presentation.request.LoginRequest;
-import com.mmp.beacon.Security.query.response.UserProfileResponse;
+import com.mmp.beacon.security.presentation.request.CreateUserRequest;
+import com.mmp.beacon.security.query.response.UserProfileResponse;
 import com.mmp.beacon.user.domain.User;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
